+++ CHANGELOG.rst
@@ -5,8 +5,7 @@ res-address changelog
 2.0.0
 -----
 
-* Breaking change: add scheme parsing, returned at the beginning of the tuple.
-* Ignore query strings.
+* Breaking change: add parsing of scheme, query string, username and password in the tuple returned.
 * Add matrix of Python versions to test on CI.